+++ src/components/Hand/index.tsx
@@ -7,12 +7,13 @@ const Hand = (): JSX.Element => {
   const { state } = React.useContext(Context);
   const { handCards, shouldShowHand } = state;
   const handleMouseDown = () => null;
+  const cards = getCards({ cardIDs: handCards, handleMouseDown, isHandCard: true });
   return (
     <Style
       cardCount={handCards.length}
       className={`hand${shouldShowHand ? '' : ' hidden'}`}
       onClick={(event: React.MouseEvent<HTMLInputElement>): void => event.stopPropagation()}>
-      {getCards({ cardIDs: handCards, handleMouseDown, isHandCard: true })}
+      {cards.length ? cards : <p className='empty'>Your hand is empty.</p>}
     </Style>
   );
 };